## Summary:
* Use the stable rust toolchain instead of a specific version (#18)
* Remove rust-toolchain

Issue: XXX-XXXX

## Test plan:

Author: jeremywiebe

Reviewers: jeremywiebe, jaredly

Required Reviewers: 

Approved by: jaredly

Checks: ✅ Run tests, ✅ Build wasm files for tests

Pull request URL: https://github.com/Khan/fastlike/pull/3

+++ xqd_body.go
@@ -9,7 +9,7 @@ func (i *Instance) xqd_body_new(handle_out int32) int32 {
 	bhid, _ := i.bodies.NewBuffer()
 	i.abilog.Printf("body_new: handle=%d", bhid)
 	i.memory.PutUint32(uint32(bhid), int64(handle_out))
-	return int32(XqdStatusOK)
+	return XqdStatusOK
 }
 
 func (i *Instance) xqd_body_write(handle int32, addr int32, size int32, body_end int32, nwritten_out int32) int32 {
@@ -19,7 +19,7 @@ func (i *Instance) xqd_body_write(handle int32, addr int32, size int32, body_end
 
 	body := i.bodies.Get(int(handle))
 	if body == nil {
-		return int32(XqdErrInvalidHandle)
+		return XqdErrInvalidHandle
 	}
 
 	// Copy size bytes starting at addr into the body handle
@@ -27,37 +27,37 @@ func (i *Instance) xqd_body_write(handle int32, addr int32, size int32, body_end
 	if err != nil {
 		// TODO: If err == EOF then there's a specific error code we can return (it means they
 		// didn't have `size` bytes in memory)
-		return int32(XqdError)
+		return XqdError
 	}
 
 	// Write out how many bytes we copied
 	i.memory.PutUint32(uint32(nwritten), int64(nwritten_out))
 
-	return int32(XqdStatusOK)
+	return XqdStatusOK
 }
 
 func (i *Instance) xqd_body_read(handle int32, addr int32, maxlen int32, nread_out int32) int32 {
 	body := i.bodies.Get(int(handle))
 	if body == nil {
-		return int32(XqdErrInvalidHandle)
+		return XqdErrInvalidHandle
 	}
 
 	buf := bytes.NewBuffer(make([]byte, 0, maxlen))
 	ncopied, err := io.Copy(buf, io.LimitReader(body, int64(maxlen)))
 	if err != nil {
 		i.abilog.Printf("body_read: error copying got=%s", err.Error())
-		return int32(XqdError)
+		return XqdError
 	}
 
 	nwritten, err2 := i.memory.WriteAt(buf.Bytes(), int64(addr))
 	if err2 != nil {
 		i.abilog.Printf("body_read: error writing got=%s", err2.Error())
-		return int32(XqdError)
+		return XqdError
 	}
 
 	if ncopied != int64(nwritten) {
 		i.abilog.Printf("body_read: error copying copied=%d wrote=%d", ncopied, nwritten)
-		return int32(XqdError)
+		return XqdError
 	}
 
 	i.abilog.Printf("body_read: handle=%d copied=%d", handle, ncopied)
@@ -65,7 +65,7 @@ func (i *Instance) xqd_body_read(handle int32, addr int32, maxlen int32, nread_o
 	// Write out how many bytes we copied
 	i.memory.PutUint32(uint32(nwritten), int64(nread_out))
 
-	return int32(XqdStatusOK)
+	return XqdStatusOK
 }
 
 func (i *Instance) xqd_body_append(dst_handle int32, src_handle int32) int32 {
@@ -73,19 +73,19 @@ func (i *Instance) xqd_body_append(dst_handle int32, src_handle int32) int32 {
 
 	dst := i.bodies.Get(int(dst_handle))
 	if dst == nil {
-		return int32(XqdErrInvalidHandle)
+		return XqdErrInvalidHandle
 	}
 
 	src := i.bodies.Get(int(src_handle))
 	if src == nil {
-		return int32(XqdErrInvalidHandle)
+		return XqdErrInvalidHandle
 	}
 
 	// replace the destination reader with a multireader that reads first from the original reader
 	// and then from the source
 	dst.reader = io.MultiReader(dst.reader, src)
 
-	return int32(XqdStatusOK)
+	return XqdStatusOK
 }
 
 func (i *Instance) xqd_body_close(handle int32) int32 {
@@ -93,5 +93,5 @@ func (i *Instance) xqd_body_close(handle int32) int32 {
 
 	i.bodies.Close(int(handle))
 
-	return int32(XqdStatusOK)
+	return XqdStatusOK
 }
